requirements.txt
# Apex Lead Gen - Requirements
praw>=7.7.0
tweepy>=4.14.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
requests>=2.31.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
